Split ContentView into multiple files
